more add command validations and authorization checks

+++ Online_Store/Commands/ProductCommands/RemoveProductCommand.cs
@@ -4,6 +4,7 @@ using Online_Store.Core.Providers;
 using Online_Store.Core.Services.User;
 using Online_Store.Data;
 using Online_Store.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -33,10 +34,13 @@ namespace Online_Store.Commands.ProductCommands
             {
                 return "You must Login First!";
             }
-
+            
             IList<string> parameters = TakeInput();
             string productName = parameters[0];
-
+            if (!this.context.Products.Any(x => x.ProductName == productName))
+            {
+                throw new ArgumentException("Product does not exist.");
+            }
             return this.productService.RemoveProductWithName(productName);
         }
